Updated favicon code

+++ bookmarks/delete_bookmark.inc.php
@@ -1,19 +1,20 @@
 <?php
 
-	require_once(dirname(__DIR__, 1) . '/header.min.php');
+	require_once(realpath(dirname(__DIR__, 1) . '/header.min.php'));
 	logged_in_only();
 
 	if (!empty($username) && $username !== 'demo') {
-debug_logger(name: 'SERVER[QUERY_STRING]', variable: $_SERVER['QUERY_STRING'], file: __FILE__, function: __FUNCTION__);
+debug_logger(name:'SERVER[QUERY_STRING]', variable:$_SERVER['QUERY_STRING'], file:__FILE__, function:__FUNCTION__);
 
 		$qs = ltrim($_SERVER['QUERY_STRING'], '?');
 		parse_str($qs, $qs_arr);
 
-		$bmlist          = $qs_arr['bmlist'];     // Comma separated string of bookmark ID's, e.g. 4186,4193,5825
-		$icons_to_delete = $qs_arr['bookmarks'];  // Array of favicon name's.
+		$icons_to_delete = $qs_arr['bookmarks'];
 
-debug_logger(name: 'bmlist', variable: $bmlist, file: __FILE__, function: __FUNCTION__);
-debug_logger(name: 'icons_to_delete', variable: $icons_to_delete, file: __FILE__, function: __FUNCTION__);
+		$bmlist = $qs_arr['bmlist'];
+
+debug_logger(name:'bmlist', variable:$bmlist, file:__FILE__, function:__FUNCTION__);
+debug_logger(name:'icons_to_delete', variable:$icons_to_delete, file:__FILE__, function:__FUNCTION__);
 
 
 		$query = sprintf("
@@ -35,26 +36,13 @@ debug_logger(name: 'icons_to_delete', variable: $icons_to_delete, file: __FILE__
 
 		foreach ($icons_to_delete as $favicon) {
 			if (! str_contains($favicon, 'bookmark.gif')) {
-				$count_query = sprintf("
-					SELECT COUNT(*)
-					FROM `obm_bookmarks`
-					WHERE `favicon` = '%d'
-					AND `user` = '%s'",
-						$mysql->escape($favicon),
-						$mysql->escape($username)
-				);
-				if ($mysql->query($count_query)) {
-					if (mysql_result($mysql->result, 0) >= 1) {
-					  // Skip deletion if more than one bookmark uses the same favicon.
-						continue;
+				if (is_file(DOC_ROOT .'/icons/'. $favicon)) {
+					if (!str_contains($favicon, 'bookmark')) {
+						unlink(DOC_ROOT .'/icons/'. $favicon);
 					}
 				}
-
-				if (is_file(DOC_ROOT .'/icons/'. $favicon)) {
-					unlink(DOC_ROOT .'/icons/'. $favicon);
-				}
 				else {
-					debug_logger(name: 'ERROR -- No favicon was found for deletion.', variable: $favicon, file: __FILE__, function: __FUNCTION__);
+					debug_logger(name:'ERROR -- No favicon was found for deletion.', variable:$favicon, file:__FILE__, function:__FUNCTION__);
 				}
 			}
 		}
@@ -62,7 +50,7 @@ debug_logger(name: 'icons_to_delete', variable: $icons_to_delete, file: __FILE__
 		echo "<script> reloadclose(); </script>";
 	}
 	else {
-		echo 'Demo users cannot delete bookmarks.<br><br>' . PHP_EOL;
+		echo 'Demo users cannot delete folders.<br><br>' . PHP_EOL;
 		echo '<input type="button" value=" Cancel " onclick="self.close()">';
 	}
